fix: reverted import syntax for files and changes model DT to INTEGER

+++ server/db/models/Cart_Item.js
@@ -5,7 +5,7 @@ const Order = require('./Order')
 
 const Cart_Item = db.define('Cart_Item', {
     total: {
-        type: Sequelize.DECIMAL(10,2),
+        type: Sequelize.INTEGER,
         defaultValue: 0
     },
     quantity: {
@@ -13,6 +13,7 @@ const Cart_Item = db.define('Cart_Item', {
         defaultValue: 0
     }
 });
+
 Order.belongsToMany(Product, { through: Cart_Item });
 Product.belongsToMany(Order, { through: Cart_Item });
